Make quasi-tautological test to be less brittle
If you read the test and the implementation, the test is basically
replicating the implementation making it very brittle, since almost any
change in the implementation will most likely break the test, even if
the output of the method doesn't change.

Instead, assert for the real expected values, which is also more
readable.